all2api/playground2api/requirements.txt
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
PyYAML>=6.0.0

